+++ src/solverpy/builder/autotune/autotune.py
@@ -8,9 +8,8 @@ import multiprocessing
 
 from ...tools import human, redirect
 from ...trains import svm
-from ...task.bar import BuilderBar
 from . import tune, build
-from ...benchmark.report import markdown
+from .listener import AutotuneListener
 
 logger = logging.getLogger(__name__)
 
@@ -49,7 +48,7 @@ def tuner(
    max_leaves=2048,
    queue=None,
 ):
-   if queue: queue.put(("TUNING", time.time()))
+   if queue: queue.put(("tuning", time.time()))
    (xs, ys) = svm.load(f_train)
    dtrain = lgb.Dataset(xs, label=ys)
    (xs0, ys0) = svm.load(f_test) if f_test != f_train else (xs, ys)
@@ -94,67 +93,16 @@ def tuner(
          best = best0 
          params.update(params0)
    
-   if queue: queue.put(("TUNED", time.time()))
+   if queue: queue.put(("tuned", time.time()))
    ret = best + (params, pos, neg)
    if queue: 
-      queue.put(("RESULT", ret))
+      queue.put(("result", (ret,)))
    else:
       return ret
 
 def prettytuner(*args, **kwargs):
-   
-   bar = None
-   desc = "trial"
-   iters = ""
-   t_start = 0
-   t_end = 0
-   f_mod = None
-   nick = None
-   it = None
-   values = None
-   table = None
-   header = None
 
-   def handle(msg):
-      nonlocal bar, desc, iters, t_start, t_end, f_mod, nick, it, values, table, header
-      (key, val) = msg
-      if key == "RESULT":
-         return val
-      elif key == "BUILD":
-         (f_mod, total) = val
-         logger.debug(f"building model: {f_mod}")
-         bar = BuilderBar(total, desc)
-      elif key == "BUILT":
-         bar.close()
-         logger.debug(f"model {f_mod} built: score={val:.4f}")
-      elif key == "ITER":
-         (n, total, loss) = val
-         bar.done(loss)
-      elif key == "TRIALS":
-         (nick, iters, timeout) = val
-         logger.info(f"Running tuning phase: {nick}")
-         iters = f"/{iters}" if iters else ""
-         header = ["it", nick, "score", "acc", "time"]
-         table = []
-      elif key == "TRY":
-         (nick, it, values) = val
-         desc = f"{nick}[{it+1}{iters}]"
-         values = ", ".join("%.4f"%v if type(v) is float else str(v) for v in values)
-         desc = f"[{it+1}{iters}] {values:8s}"
-         #logger.info(f"Starting trial: {desc}: ({values})")
-      elif key == "TRIED":
-         table.append((it, values, f"{val[0]:.4f}", human.humanacc(val[1]), human.humantime(val[2])))
-         #logger.info(f"Trial result: {val[0]:.4f} ; {human.humanacc(val[1])} ; {human.humantime(val[2])}")
-      elif key == "TRIALED":
-         table = [""] + markdown.table(header, table) + [""]
-         logger.info(markdown.dump(table, prefix="> "))
-      elif key == "TUNING":
-         t_start = val
-      elif key == "TUNED":
-         t_end = val
-      else:
-         print(msg)
-      return None
+   listener = AutotuneListener()
 
    queue = multiprocessing.Queue()
    kwargs["queue"] = queue
@@ -166,7 +114,7 @@ def prettytuner(*args, **kwargs):
       p.start()
       while True:
          msg = queue.get()
-         result = handle(msg)
+         result = listener.listen(msg)
          if result: 
             break
    except (Exception, KeyboardInterrupt) as e: